% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasr.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the map of incidences occuring in the state and the year set by the use}
\usage{
fars_map_state(state.num, year = c(2013, 2014, 2015))
}
\arguments{
\item{state.num}{state number}

\item{year}{Year number}
}
\value{
A dataframe summarizing number of months appearing in the fars data by year
}
\description{
Plot the map of incidences occuring in the state and the year set by the use
}
\details{
The function may return errors when any of the years cannot be found in the filename.
Or the stat.num is invalid.
}
\examples{
library(dplyr)
library(maps)
library(farc)
fars_map_state(1, 2013)

}
