% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasr.R
\name{fars_read}
\alias{fars_read}
\title{Add together two numbers}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of a csv file.}
}
\value{
A tibble.
}
\description{
Add together two numbers
}
\details{
The function may return errors when the filename cannot be found
}
\examples{
fas_read('fasdata.csv')
}
