% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasr.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in data and make sure years in the data are consistent with their fileanmes}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A string of year numbers}
}
\value{
A list of dataframes. Each data frame contains two columns, month and year
}
\description{
Read in data and make sure years in the data are consistent with their fileanmes
}
\details{
The function may return errors when any of the years cannot be found in the filename
}
\examples{
library(dplyr)
make_filename(2013:2015)
}
