% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasr.R
\name{make_filename}
\alias{make_filename}
\title{Makde filenames in a format of accident_YY.csv.bz2}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year number}
}
\value{
A character "accident_year.csv.bz2" where year is the input
}
\description{
Makde filenames in a format of accident_YY.csv.bz2
}
\details{
The function may return errors when year cannot be converted to an integer
}
\examples{
make_filename(year)
}
