% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasr.R
\name{fars_read}
\alias{fars_read}
\title{Read a fars file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of a csv file.}
}
\value{
A tibble.
}
\description{
Read a fars file
}
\details{
The function may return errors when the filename cannot be found
}
\examples{
library(farc)
library(dplyr)
system.file("extdata", 'accident_2013.csv.bz2' , package = "farc") \%>\% fars_read()

}
