% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasr.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize number of months appearing in the fars data by year}
\usage{
fars_summarize_years(years = c(2013, 2014, 2015))
}
\arguments{
\item{years}{A string of year numbers}
}
\value{
A dataframe summarizing number of months appearing in the fars data by year
}
\description{
Summarize number of months appearing in the fars data by year
}
\examples{
library(dplyr)
library(tidyr)
library(farc)
fars_summarize_years(2013:2014)
}
